Introduced Trial and Error!
The code solves the sudoku to the extent it can and then starts trial
and error. Goes to each cell which has only two options and fills both
of them one by one and solves the rest of the sudoku normally. If the
score ever reaches 81, the program stops. If all cells with only  two
alternatives are checked and still no solution is found, I give up.